Added generic card for Tasks
