+++ src/components/challenges/active.tsx
@@ -1,7 +1,4 @@
 import { FC, useState } from 'react';
-// import classNames from '@/components/challenges/styles.module.scss';
-// import Details from '@/components/challenges/details';
-// import Participants from '@/components/challenges/participants';
 import Card from '@/components/Card/index';
 
 type ActiveProps = {
@@ -34,10 +31,6 @@ type ActiveProps = {
 const Active: FC<ActiveProps> = ({ content }) => {
   const [isUserSubscribed, setUserSubscribed] = useState(content.is_user_subscribed);
 
-  // const subscribeEventHandler = async () => {
-  //   setUserSubscribed(1);
-  // };
-
   const taskData = {
     Level: content.level,
     Challenge_Started: content.start_date,
@@ -61,10 +54,8 @@ const Active: FC<ActiveProps> = ({ content }) => {
     // eslint-disable-next-line array-callback-return
     users.map((user) => {
       participants.push({
-        firstName: user.first_name,
-        lastName: user.last_name,
         userName: user.rds_member_id,
-        imgUrl: user.img,
+        imgUrl: `${process.env.NEXT_PUBLIC_GITHUB_IMAGE_URL}${user.rds_member_id}/img.png`,
         key: user.rds_member_id,
       });
     });
@@ -88,33 +79,6 @@ const Active: FC<ActiveProps> = ({ content }) => {
       }
       key={content.title}
     />
-    // <div className={classNames.boxContent}>
-    //   <p className={classNames.heading}>{content.title}</p>
-    //   <Details text="Level" value={content.level} />
-    //   <Details text="Challenge Started" value={content.start_date} />
-    //   <Details text="Challenge Ends" value={content.end_date} />
-    //   <div className={classNames.participants}>
-    //     <Details
-    //       text="Active Participants"
-    //       value={content.participants.length}
-    //     />
-    //     <Participants participants={content.participants} />
-    //   </div>
-    //   {
-    //     !isUserSubscribed && (
-    //       <p className={classNames.activeBtn}>
-    //         <button
-    //           onClick={subscribeEventHandler}
-    //           onKeyDown={subscribeEventHandler}
-    //           tabIndex={0}
-    //           type="button"
-    //         >
-    //           I will do this
-    //         </button>
-    //       </p>
-    //     )
-    //   }
-    // </div>
   );
 };
