+++ src/components/pullRequests/index.tsx
@@ -1,7 +1,7 @@
-import { FC, useState } from 'react';
-import classNames from '@/components/pullRequests/pullRequests.module.scss';
+import { FC } from 'react';
+import Card from '@/components/Card/index';
 
-interface prData {
+interface pullRequestType {
   title: string;
   username: string;
   createdAt: string;
@@ -9,7 +9,7 @@ interface prData {
   url: string;
 }
 
-const pullRequest: FC<prData> = ({
+const pullRequest: FC<pullRequestType> = ({
   title,
   username,
   createdAt,
@@ -19,68 +19,34 @@ const pullRequest: FC<prData> = ({
   const created = new Date(createdAt);
   const updated = new Date(updatedAt);
   const presentDate = new Date();
+
   const createdAgo = `${Math.ceil(
     (presentDate.getTime() - created.getTime()) / (1000 * 3600 * 24),
   )}days ago`;
   const updatedAgo = `${Math.ceil(
     (presentDate.getTime() - updated.getTime()) / (1000 * 3600 * 24),
   )}days ago`;
-  const [createdTime, updateCreatedTime] = useState(createdAgo);
-  const [updatedTime, updateUpdatedTime] = useState(updatedAgo);
+
+  const prData = {
+    CreatedBy: username,
+    Created: createdAgo,
+    Updated: updatedAgo,
+  };
+  const PR:any[] = [];
+  function getPr() {
+    // eslint-disable-next-line array-callback-return
+    Object.entries(prData).map(([key, value]) => {
+      PR.push({ key, value });
+    });
+    return PR;
+  }
 
   return (
-    <div className={classNames.pullRequest}>
-      <div className={classNames.prTitle}>
-        <span className={classNames.title}>Title:</span>
-        {title}
-      </div>
-      <div className={classNames.statusLable}>
-        Created by:
-        {' '}
-        <span className={classNames.infoEl}>{username}</span>
-      </div>
-      <div className={classNames.statusLable}>
-        Created:
-        {' '}
-        <span
-          className={classNames.infoEl}
-          onMouseEnter={() => {
-            updateCreatedTime(created.toLocaleString());
-          }}
-          onMouseLeave={() => {
-            updateCreatedTime(createdAgo);
-          }}
-        >
-          {createdTime}
-        </span>
-      </div>
-      <div className={classNames.statusLable}>
-        Updated:
-        {' '}
-        <span
-          className={classNames.infoEl}
-          onMouseEnter={() => {
-            updateUpdatedTime(updated.toLocaleString());
-          }}
-          onMouseLeave={() => {
-            updateUpdatedTime(updatedAgo);
-          }}
-        >
-          {updatedTime}
-        </span>
-      </div>
-      <div className={classNames.linkCt}>
-        <a
-          className={classNames.prLink}
-          href={url}
-          target="_blank"
-          rel="noreferrer"
-        >
-          Open PR in Github
-        </a>
-      </div>
-    </div>
+    <Card
+      title={{ text: title, link: url }}
+      data={getPr()}
+      key={title}
+    />
   );
 };
-
 export default pullRequest;
