+++ src/pages/openPRs.tsx
@@ -1,7 +1,7 @@
 import { FC, useState, useEffect } from 'react';
 import Head from '@/components/head';
 import Layout from '@/components/Layout';
-import PullRequest from '@/components/pullRequests/indexModified';
+import PullRequest from '@/components/pullRequests';
 import fetch from '@/helperFunctions/fetch';
 import CardShimmer from '@/components/Loaders/cardShimmer';
